refactor: split mode command for user mode (not yet implemanted) and channel mode
