+++ test/commands/mode/channel.test.cpp
@@ -1,0 +1,3 @@
+#include "catch.hpp"
+
+#include "../src/commands/mode/channel.cpp"
