+++ test/commands/mode/channel/invite_only.test.cpp
@@ -1,0 +1,3 @@
+#include "catch.hpp"
+
+#include "../src/commands/mode/channel/invite_only.cpp"
