+++ test/commands/mode/channel/operator.test.cpp
@@ -1,0 +1,3 @@
+#include "catch.hpp"
+
+#include "../src/commands/mode/channel/operator.cpp"
